pyproject.toml
[project]
name = "gesture-ml"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = ["numpy>=1.24","torch>=2.0","mediapipe==0.10.14","opencv-python>=4.8","fastapi>=0.100","uvicorn>=0.23","scikit-learn>=1.3","pyyaml>=6.0","tqdm>=4.65","matplotlib>=3.7"]
[project.scripts]
gesture-train = "src.train:main"
gesture-predict = "src.predict:demo"